requirements.txt
transformers==3.0
jsonnet~=0.14.0
nltk~=3.4
networkx~=2.2
entmax
pyrsistent~=0.14.9
bpemb~=0.2.11
stanford-corenlp~=3.9.2
torchtext~=0.3.1
flask~=2.2.5
attrs~=23.1.0
asdl~=0.1.5
protobuf==3.20.*
astor~=0.8.1
flask_cors~=3.0.10
Snowflake-connector-python~=3.0.3